Third arg INSTALLDIR to apertium-gen-modes now gives the data prefix
So you can for example have apertium installed to prefix /usr, but
install language pairs to $HOME/local. Backwards compatible with old
makefiles (the one and two argument calls), but now your makefile can
also have

	apertium-gen-modes modes.xml --to $(apertium_isl_engdir)

to say that installed datafiles are in that specific directory.

Puts installable modes in ./*.mode and all dev modes in modes/*.mode

Cleanup as well: moved most of the script out of the makefile and into
apertium-gen-modes.in, quoted bash strings etc.